+++ NEWS.md
@@ -6,6 +6,10 @@
   * exclusions for less severer conditions when more severer conditions are flagged
   * Improved POA, NPOA, and EXEMPTPOA.  This came about from #20.
 
+* Make internal data.frame tool `mdcr_duplicated` data.table aware.
+
+* Elixhauser (Quan 2005) - added missing ICD-10 codes to the mappings
+
 ## New Features
 
 * `comorbidities()` will return a `tibble` with the input data is a `tbl_df`